Tue, May 10, 2022, 11:03 AM +00:00

+++ sentimental-credit/credit.py
@@ -17,7 +17,7 @@ luhn = 0
 
 for i in range(17):
     if i % 2 == 0 and (((card % 10) * 2) > 9):
-        luhn += (((card % 10) * 2) % 10) + ((((card % 10) * 2) / 10) % 10)
+        luhn += (((card % 10) * 2) % 10) + (int(((card % 10) * 2) / 10) % 10)
     elif (i % 2 > 0):
         luhn += (card % 10)
     else:
